pyproject.toml
[tool.poetry]
name = "worker-sample"
version = "0.1.0"
description = "KuFlow worker example in Python for your process"
authors = ["KuFlow S.L. <kuflow@kuflow.com>"]
readme = "README.md"
packages = [{include = "worker_sample"}]

[tool.poetry.dependencies]
python = "^3.8"
kuflow-temporal-activity-kuflow = "^0.4.1"
pyyaml = "^6.0"
pillow = "^9.4.0"
pytesseract = "^0.3.10"

[tool.poetry.dev-dependencies]
ipython = "^7.29.0"
flake8 = "4.0.1"
black = "^23.1.0"
pytest = "^7.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"
testpaths = ["tests"]
python_files= "test_*.py"
